feat: Solved the problem of the post-configuration optimization interface uninstalling the package resulting in a dead card
解决后配置优化界面卸载包命令导致的卡死问题

Log:
  1. 优化代码结构，增加类RemovePackageModel用于管理包卸载
  2. 优化界面显示，增加卸载包失败的日志显示
  3. 优化界面操作，如果卸载包失败，则需要用户点击确认后再进入系统
  4. 增加system_module_debug用于控制调试的模块是否发布
  5. 增加system_timeout用于控制系统中超时时间，当前用于卸载包超时时间

Task: https://pms.uniontech.com/zentao/task-view-59195.html

+++ hooks/first_boot_setup.sh
@@ -79,8 +79,14 @@ uninstall_packages() {
   add_uninstall_package "imagemagick*"
 
   local list=$(installer_get "DI_UNINSTALL_PACKAGES")
-  apt-get -y purge ${list}
-  # apt autoremove -y   // 屏蔽代码解决磁盘加密时选择非图形组建安装系统成功后，系统无法重启问题。根因：由于cryptsetup被apt autoremove 卸载导致无法启动
+
+  # 增加system_module_debug判断，用于隔离安装包超时逻辑的修改
+  local module=$(installer_get "system_module_debug")
+  if [ "x$module" != "xtrue" ]; then
+    apt-get -y purge ${list}
+    # apt autoremove -y   // 屏蔽代码解决磁盘加密时选择非图形组建安装系统成功后，系统无法重启问题。根因：由于cryptsetup被apt autoremove 卸载导致无法启动
+  fi
+
 }
 
 # Replace lightdm.conf with lightdm.conf.real.
